requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow>=2.10.0
pydantic==2.5.0
numpy>=1.21.0
opencv-python==4.8.1.78
pandas>=1.5.0
ndjson==0.3.1
scikit-learn>=1.3.0
seaborn>=0.12.0
pillow>=9.5.0
python-multipart==0.0.6
matplotlib>=3.6.0